move/rename obro to Meridian RGB
