+++ keyboards/primekb/meridian_rgb/info.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "Obro",
+    "keyboard_name": "Meridian RGB",
     "url": "https://www.primekb.com",
     "maintainer": "holtenc",
     "layouts": {
